fix(sqlite): read the three measurement columns as float64
Widens the read model and the JSON contract ahead of the write-path
change, so a fractional value can be read back before one can be
written. Reversed, the current-observation endpoint would break between
commits.

Summary.LightningTotal is the site that fails silently: SUM over a REAL
column returns REAL, but SUM(2,3) = 5.0 still scans into NullInt64 as 5,
so every existing summary test passes either way. Only a fractional sum
errors -- hence the new test seeding 2.5.

writer_test.go was a silent survivor of the same kind: it declared the
columns as plain int64 and its integral fixtures kept scanning fine from
a REAL column, so it stayed green while asserting the wrong type.

The wire change is byte-identical for integral values -- Go marshals
float64(5) as 5, not 5.0. precip_type stays an integer end to end.

+++ internal/httpserver/observations_test.go
@@ -64,12 +64,12 @@ func testDepsWithObservations(reader ObservationReader) Deps {
 // one derived (server-computed) field, and 404s with the sentinel error.
 func TestAPI_CurrentObservation(t *testing.T) {
 	t.Run("returns_contract_c_shape", func(t *testing.T) {
-		windSample := int64(3)
+		windSample := 3.0
 		precip := int64(1)
 		lightningDist := 2.1
-		lightningCount := int64(4)
+		lightningCount := 4.0
 		battery := 3.6
-		reportInterval := int64(5)
+		reportInterval := 5.0
 
 		reader := &fakeObservationReader{
 			obs: sqlite.Observation{
@@ -322,7 +322,7 @@ func TestHandleSummary_OK(t *testing.T) {
 		WindMax:        sql.NullFloat64{Float64: 8, Valid: true},
 		GustMax:        sql.NullFloat64{Float64: 12, Valid: true},
 		RainTotal:      sql.NullFloat64{Float64: 3.5, Valid: true},
-		LightningTotal: sql.NullInt64{Int64: 4, Valid: true},
+		LightningTotal: sql.NullFloat64{Float64: 4, Valid: true},
 	}}
 
 	srv := New(testDepsWithObservations(reader))
